refactor convert code so it can be used outside utils
